Clean up routes.js
